Reworked Product class
