+++ app.py
@@ -1,16 +1,15 @@
+import webbrowser
 from flask import Flask, render_template, url_for, redirect, request, session, flash
 from flask_sqlalchemy import SQLAlchemy
 
-from check_stock import check_stock, Product, validate_url
+from web_helper import is_url_valid
+from product import Product
 
+ENV = "dev"
 
 app = Flask(__name__)  # reference this file
-app.secret_key = "\xdb\xf5xn-\xaa\xf4\xdeHw\xacc\xb9\xc8\xcdA\xfe\xcfxT\xe4\xf3\xe4\x89"
-# app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///products.sqlite3"
-app.config["SQLALCHEMY_DATABASE_URI"] = "postgres://ptbmzeoojhjyqn:183d0bd5d83037665ee999566150fb2dc8af79939be350c2db10d6249546af05@ec2-184-73-243-101.compute-1.amazonaws.com:5432/d1sldsv1g5i1pq"
-app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-
 db = SQLAlchemy(app)
+
 
 class Products(db.Model):
     __tablename__ = "Products"
@@ -19,22 +18,22 @@ class Products(db.Model):
     site_name = db.Column("site_name", db.String)
 
     def __init__(self, product):
-        self.url = product.get_url()
-        self.product_name = product.get_product_name()
-        self.site_name = product.get_site_name()
+        self.url = product.url
+        self.product_name = product.product_name
+        self.site_name = product.site_name
 
 
 @app.route("/")
 def index():
-    return render_template("index.html", values=Products.query.all(), check_stock=check_stock)
+    return render_template("index.html", values=Products.query.all(), check_stock=Product.check_stock)
 
 
-@app.route("/products_list", methods=["POST", "GET"])
+@ app.route("/products_list", methods=["POST", "GET"])
 def products():
     if request.method == "POST":
         if "add_url" in request.form:
             url = request.form["add_url"]
-            if validate_url(url):
+            if is_url_valid(url):
                 # if link is not already in the database then add it
                 if not Products.query.filter_by(url=url).first():
                     product = Product(url)
@@ -51,4 +50,15 @@ def products():
 
 
 if __name__ == "__main__":
+    app.secret_key = "\xdb\xf5xn-\xaa\xf4\xdeHw\xacc\xb9\xc8\xcdA\xfe\xcfxT\xe4\xf3\xe4\x89"
+    # app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+    if ENV == "prod":
+        app.config["SQLALCHEMY_DATABASE_URI"] = "postgres://ptbmzeoojhjyqn:183d0bd5d83037665ee999566150fb2dc8af79939be350c2db10d6249546af05@ec2-184-73-243-101.compute-1.amazonaws.com:5432/d1sldsv1g5i1pq"
+        app.debug = False
+    elif ENV == "dev":
+        app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///products.sqlite3"
+        app.debug = True
+        webbrowser.open("http://127.0.0.1:5000/")
+        db.create_all()
+
     app.run()
